requirements.txt
flask
sentence-transformers
faiss-cpu
chardet
pdfplumber
orjson
